refactor: now() 并入 crossPlatform，去掉单函数文件 src/timing.ts
now() 本质是平台时钟，与 sdk()/getPerformanceManager() 同类，无需独立文件。
测试改为直接 spyOn(crossPlatform,'now') 控制时钟，比经 sdk().getPerformance()
更直接，framerate 测试也不再需要 mock sdk。行为不变，466 测试全过。
